+++ src/app/(header)/create/_components/CreatePlanStepper/StepButtonGroup/index.scss
