+++ src/app/(header)/create/_components/CreatePlanStepper/index.scss
