+++ src/app/(header)/create/index.scss
@@ -1,25 +1,11 @@
-.create-page {
-  padding: 1.75rem 2.5rem;
-  // height: 100%;
+.new-create-page{
+  width: 100%;
+  height: 100%;
   display: flex;
   flex-direction: column;
-  // overflow: auto;
+  align-items: center;
 
-  &__remind {
-    &:before {
-      content: '';
-      display: block;
-      height: 0.3px;
-      background-color: var(--origin-gray-100);
-      border: 1px solid var(--origin-gray-100);
-      margin: 2rem 0 2rem 0;
-    }
-  }
-
-  &__button__container {
-    margin-top: 2rem;
-    display: flex;
-    justify-content: center;
-    gap: 13rem;
+  &__title{
+    margin-top: 0rem;
   }
 }
