--- src/app/(header)/newCreate/index.scss
@@ -1,11 +1,0 @@
-.new-create-page{
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  &__title{
-    margin-top: 0rem;
-  }
-}
